Fix finding of patterns and linetypes from measurement system.
The current code depends on the document unit instead of the measurement system - although the user interface says otherwise. In addition, the linetypes are initialized before the document has its variables defined.
Fixed the CMakeLists.txt to be consisted with qmake.
Fixed the plugin loaded to not expect _debug in dll plugin names - to be consistent with changes to both qmake and cmake.